Move filter helpers to lib and reuse ownership in selection dialog
