+++ src/lib/categoryFilterUtils.ts
