+++ src/lib/filterOwnership.ts
